Sample code for retrieving AAD user object ID
